Added player registration

+++ src/main/kotlin/uk/co/scarfebread/wizardbeast/model/Player.kt
@@ -3,4 +3,4 @@ package uk.co.scarfebread.wizardbeast.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Player(val id: String, val name: String, val x: Int, val y: Int)
+data class Player(val id: String, val name: String, val x: Float, val y: Float)
